CRUD outfits in OutfitStore via UseCases
