+++ OOTD/UseCases/Outfit/AddOutfits.swift
@@ -1,0 +1,16 @@
+//
+//  AddOutfits.swift
+//  OOTD
+//
+//  Created by Hiroshi Matsui on 2024/11/27.
+//
+
+import Foundation
+
+struct AddOutfits {
+    let repository: OutfitRepository
+
+    func callAsFunction(_ outfits: [Outfit]) async throws {
+        try await repository.create(outfits)
+    }
+}
